refactor: split the strategies into dedicated modules, with better naming
