+++ src/core/strategies/sign.js
