+++ src/core/strategies/suffix.js
@@ -20,13 +20,6 @@ import { terms } from '../terms.js';
 import tokensHelper from '../tokens.js';
 
 /**
- * Adds a multiply operator before the value.
- * @param {string} value - The value to modify.
- * @returns {string} - The modified value.
- */
-const multiplyBefore = value => `*${value}`;
-
-/**
  * Adds a multiply operator after the value.
  * @param {string} value - The value to modify.
  * @returns {string} - The modified value.
@@ -34,97 +27,11 @@ const multiplyBefore = value => `*${value}`;
 const multiplyAfter = value => `${value}*`;
 
 /**
- * Adds a space before the value.
- * @param {string} value - The value to modify.
- * @returns {string} - The modified value.
- */
-const spaceBefore = value => ` ${value}`;
-
-/**
  * Adds a space after the value.
  * @param {string} value - The value to modify.
  * @returns {string} - The modified value.
  */
 const spaceAfter = value => `${value} `;
-
-/**
- * List of strategies to apply for glueing tokens together with a prefix.
- * @type {valueStrategy[]}
- */
-export const prefixStrategies = [
-    {
-        // adding an opening parenthesis after a value, a unary operator, or a closing parenthesis
-        predicate(previous, next) {
-            const previousTerm = terms[previous];
-            return (
-                next === 'LPAR' &&
-                (previous === 'RPAR' ||
-                    tokensHelper.isValue(previousTerm) ||
-                    tokensHelper.isUnaryOperator(previousTerm))
-            );
-        },
-        action: multiplyBefore
-    },
-    {
-        // adding an identifier or a value after a closing parenthesis
-        predicate(previous, next) {
-            const nextTerm = terms[next];
-            return (
-                previous === 'RPAR' &&
-                nextTerm.exponent !== 'left' &&
-                (tokensHelper.isValue(nextTerm) || tokensHelper.isFunction(nextTerm))
-            );
-        },
-        action: multiplyBefore
-    },
-    {
-        // adding an identifier after a value or a unary operator
-        predicate(previous, next) {
-            const previousTerm = terms[previous];
-            const nextTerm = terms[next];
-            return (
-                (tokensHelper.isValue(previousTerm) || tokensHelper.isUnaryOperator(previousTerm)) &&
-                tokensHelper.isIdentifier(nextTerm) &&
-                nextTerm.exponent !== 'left'
-            );
-        },
-        action: multiplyBefore
-    },
-    {
-        // adding a digit after an identifier that is not a function
-        predicate(previous, next) {
-            const previousTerm = terms[previous];
-            const nextTerm = terms[next];
-            return (
-                tokensHelper.isIdentifier(previousTerm) &&
-                !tokensHelper.isFunction(previousTerm) &&
-                tokensHelper.isDigit(nextTerm)
-            );
-        },
-        action: multiplyBefore
-    },
-    {
-        // adding a value after a unary operator
-        predicate(previous, next) {
-            const previousTerm = terms[previous];
-            const nextTerm = terms[next];
-            return tokensHelper.isUnaryOperator(previousTerm) && tokensHelper.isValue(nextTerm);
-        },
-        action: multiplyBefore
-    },
-    {
-        // adding an identifier or a value after a function
-        predicate(previous, next) {
-            const previousTerm = terms[previous];
-            const nextTerm = terms[next];
-            return (
-                tokensHelper.isFunction(previousTerm) &&
-                (tokensHelper.isIdentifier(nextTerm) || !tokensHelper.isSeparator(nextTerm))
-            );
-        },
-        action: spaceBefore
-    }
-];
 
 /**
  * List of strategies to apply for glueing tokens together with a suffix.
