+++ src/core/strategies/test/prefix.spec.js
@@ -17,7 +17,7 @@
  */
 
 import { applyValueStrategies } from '../helpers.js';
-import { prefixStrategies, suffixStrategies } from '../value.js';
+import { prefixStrategies } from '../prefix.js';
 
 describe('prefixStrategies', () => {
     it('is a collection', () => {
@@ -50,39 +50,5 @@ describe('prefixStrategies', () => {
         ['(', 'COS', 'LPAR', '(']
     ])('adds a separator to %s between %s and %s', (value, previous, next, expected) => {
         expect(applyValueStrategies(value, previous, next, prefixStrategies)).toEqual(expected);
-    });
-});
-
-describe('suffixStrategies', () => {
-    it('is a collection', () => {
-        expect(suffixStrategies).toEqual(expect.any(Array));
-        expect(suffixStrategies[0]).toEqual(expect.any(Object));
-        expect(suffixStrategies[0].predicate).toEqual(expect.any(Function));
-        expect(suffixStrategies[0].action).toEqual(expect.any(Function));
-    });
-
-    it.each([
-        // value, name, next, expected
-        [')', 'RPAR', 'LPAR', ')*'],
-        [')', 'RPAR', 'NUM1', ')*'],
-        [')', 'RPAR', 'COS', ')*'],
-        ['1', 'NUM1', 'LPAR', '1*'],
-        ['!', 'FAC', 'LPAR', '!*'],
-        ['!', 'FAC', 'NUM1', '!*'],
-        ['!', 'FAC', 'COS', '!*'],
-        ['!', 'FAC', 'PI', '!*'],
-        ['!', 'FAC', 'ADD', '!'],
-        ['PI', 'PI', 'LPAR', 'PI*'],
-        ['cos', 'COS', 'LPAR', 'cos'],
-        ['PI', 'PI', 'NUM1', 'PI*'],
-        ['PI', 'PI', 'ADD', 'PI'],
-        ['cos', 'COS', 'NUM1', 'cos '],
-        ['cos', 'COS', 'ADD', 'cos'],
-        ['1', 'NUM1', 'COS', '1*'],
-        ['1', 'NUM1', 'ADD', '1'],
-        ['cos', 'COS', 'ADD', 'cos'],
-        ['cos', 'COS', 'NUM', 'cos ']
-    ])('adds a separator to %s between %s and %s', (value, previous, next, expected) => {
-        expect(applyValueStrategies(value, previous, next, suffixStrategies)).toEqual(expected);
     });
 });
